Casting to f32 not required

+++ src/rectangle_color_context.rs
@@ -77,7 +77,7 @@ impl<'a> Fill<'a> for RectangleColorContext<'a> {
             let rect = self.rect.get();
             let rect: [f32, ..4] = [rect[0] as f32, rect[1] as f32, rect[2] as f32, rect[3] as f32];
             let color = self.color.get();
-            let color: [f32, ..4] = [color[0] as f32, color[1] as f32, color[2] as f32, color[3] as f32];
+            let color: [f32, ..4] = [color[0], color[1], color[2], color[3]];
             back_end.tri_list_xy_rgba_f32(
                 rect_tri_list_xy_f32(rect),
                 rect_tri_list_rgba_f32(color)
